Started UI/Hud, fixing up AI/Pathfinding and attack loop. Few more placeholder sounds.
Fixed up direction for monsters, as they were chasing/attacking in weird directions. They now seem to move and attack correctly, knights are a bit too aggressive though possibly.

Started adding interactive doors with built in crusher functionality. Started a 3D UI/Hud which will be used for world interaction and makes it closer to being VR-ready.

Added a couple of fonts to play with.