Knights now stop aiming after push_forward
Knights no longer constantly aim, so attacks will not track the player for the entire duration. They'll now stop tracking after jumping and starting the push forward, and then land next to the Player if they move out of the way. Previously they'd track and always hit the player in close range.