Damage knockback and bounce pads added!
Enemies will bounce back based on the direction the damage came from. Will be adding in-fighting and a scaled knockback based on damage, as well as pain/flinch reaction chance etc.

Bounce pads added! They're already really fun. Currently jumps the character if they move onto it from the ground, then will bounce them the full height.

Started to switch the Player into a "target" as a start to adding monster in-fighting. Will need to add a Source to hits taken in case in-fighting is to occur.