SLOPES FIXED YAY
Using a RayCast for foot collision doesn't take into account normals, which it doesn't need to if I'm just detecting relative floor height. This means it won't cause issues when going uphill, and also functions to "mount" edges and stairs.

Had to add 1 unit of downward gravity to stop jitter when stationary. Now works perfectly.

Enemy attack angle includes the up/down angle. This means when it was at 5, it wasn't allowing the enemy to attack up slopes. Setting to 20+ fixes this.