Cleaned up is_instance_valid checks. Blood effect added to knight attacks.
Made sure that if a target instance is not valid, it assigns the player as the target. Previously it was  returning or skipping code if the instance was not valid, which was causing weird behaviour where they'd lock up in certain frozen states without a target.

Making sure that target is set to player  if the instance isn't valid makes sure the enemy code continues to run correctly.

Blood effect added to knight attacks to show that they did damage to a target. Will increase the blood amount later. Likely as part of the scaled blood splatter update.

esc toggles mouse mode, alt-q quits.

temporarily fixed an issue with mouse sensitivity not being set correctly. Likely an issue with the export variable not being read in the right order at runtime.