UI Update, Shield, projectile deflection and item lightsadded.
UI now shows shield meter and ammo amount. Going to clean up the code and just go for simple connections in future.

Added shields which let you not take damage and they'll also reverse their direction back at the attacker. It goes over cats, but that adds an interesting effect as it will then hit the enemy behind it and start an infight.

Shields can be used until they reach 0. If it hits 0, it'll need to wait until it returns to full before they can be used again. Kind of an overcharged state.

 Added item lights to help make them pop out from the scenery.